backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
pyyaml==6.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
alembic==1.12.1

